+++ lib/boolean_validation.dart
@@ -10,10 +10,13 @@ export 'src/validation_groups/validation_common.dart';
 
 export 'src/enum/email_domains.dart';
 
-export 'src/regex/regex_mobile.dart';
+export 'src/regex/mobile/regex_mobile.dart';
 export 'src/enum/country_phone_codes.dart';
 
-export 'src/regex/regex_alpha_lang.dart';
+export 'src/regex/langs/regex_alpha_lang.dart';
+export 'src/regex/langs/regex_nums.dart';
+export 'src/regex/langs/regex_punctuation.dart';
+export 'src/regex/langs/regex_script_component.dart';
 export 'src/enum/supported_languages.dart';
 
 export 'src/regex/regex_patterns.dart';
@@ -23,5 +26,6 @@ export 'src/validation_messages/validation_messages.dart';
 export 'src/validation_messages/messages_provider.dart';
 
 export 'src/core/extensions/string_extension.dart';
+export 'src/core/extensions/boolean_extension.dart';
 
 export 'src/core/common_libs.dart';